+++ changelog.md
@@ -3,6 +3,8 @@
 This project adheres to semantic versioning and all notable
 changes will be documented in this file.
 
+## [unreleased]
+
 ## [0.11.0] 2024-08-06
 
 - Disconnect client if Server.maxQoS is exceeded